Change the login to Garmin Connect
The ticket is now extracted from the login response.
The cookies are not needed anymore, as well as the
voodoo calls to modern and legacy session.

For the post_auth call I chose just an URL that worked when
analyzing the HTTP traffic in the browser, maybe there are better
ones (with less payload)?

Issue moderation/garmin-connect-export#10