Add filtering by human-readable activity type
Something that bothered me every time I downloaded a new batch of activities with this script is that it requires numeric type IDs that can be pretty hard to discover (basically same as #106) or remember.

This PR adds ability to filter out by human-readable activity type instead, so you can use `-tf walking,hiking` instead of `-tf 3,9`.

The documentation was switched to show just the human-readable filter as I think it will be preferable in vast majority of usecases, but the old ID-based filter is kept in code too for backward compatibility.